User can set reporting level
By using the `--severity [1-5]` the user can indicate what level of
issues to report on (1 - all, 5 - only the most severe).

Thus this can be used to ensure the user isn't overwhelmed with
feedback and allow the user to customize their level of failure
for CI jobs.

Related to issues #79 and #41.